backend/requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.3
bidict==0.23.1
black==25.11.0
boto3==1.41.3
botocore==1.41.3
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
flake8==7.3.0
fsspec==2025.12.0
gotrue==2.12.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
jmespath==1.0.1
jq==1.10.0
librt==0.7.3
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
motor==3.3.1
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.5
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
postgrest==0.18.0
propcache==0.4.1
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.5
pyroaring==1.0.3
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.15.0
pytokens==0.3.0
pytz==2025.2
realtime==2.27.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.15.0
s5cmd==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
storage3==0.9.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.10.0
supafunc==0.7.0
tenacity==9.1.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
